Retry when there are activities remaining
This is achieved by re-fetching and re-parsing the DOM every time
`processActivities` is called, instead of using `document`.

This commit also introduces a check for testing if the script is being
loaded in `/course/view.php` as opposed to any other path. In case the
check fails, a notification is shown to the user explaining the error.